raising errors and changing print statements
An extra part of recent_show_link was added as it was braking if the show was last month. This shouldn't matter when the code is intended to be used during the time of the show. But is nice to show the code still works when not at the show with this change too.

Some nicer print statements were added.